Moved SLD creation to its own service and allowed retrieving of layer states to better facilitate map legends.
