+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@auscope/portal-core-ui",
-  "version": "1.0.61",
+  "version": "1.0.62",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
