+++ projects/portal-core-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@auscope/portal-core-ui",
-  "version": "1.0.62",
+  "version": "1.0.63",
   "license": "(LGPL OR Apache-2.0)",
   "repository": {
     "type": "git",
