+++ projects/portal-core-ui/src/lib/service/wms/cs-wms.service.ts
@@ -10,8 +10,6 @@ import { HttpClient, HttpHeaders, HttpParams, HttpResponse } from '@angular/comm
 import { Constants, ResourceType } from '../../utility/constants.service';
 import { UtilitiesService } from '../../utility/utilities.service';
 import { RenderStatusService } from '../cesium-map/renderstatus/render-status.service';
-import { MinTenemStyleService } from '../style/wms/min-tenem-style.service';
-import { GSML41StyleService } from '../style/wms/gsml41-style.service';
 import { CQLService } from '../cql/cql.service';
 import { MapsManagerService, AcMapComponent } from '@auscope/angular-cesium';
 import { WebMapServiceImageryProvider, ImageryLayer, Resource, Rectangle } from 'cesium';
@@ -23,6 +21,7 @@ import intersect from '@turf/intersect';
 
 import * as when from 'when';
 import TileProviderError from 'cesium/Source/Core/TileProviderError';
+import { SldService } from '../style/wms/sld.service';
 
 export class ErrorPayload {
   constructor(
@@ -60,6 +59,7 @@ export class CsWMSService {
     private mapsManagerService: MapsManagerService,
     private layerStatusService: LayerStatusService,
     private deviceService: DeviceDetectorService,
+    private sldService: SldService,
     @Inject('env') private env,
     @Inject('conf') private conf
   ) { }
@@ -70,13 +70,12 @@ export class CsWMSService {
   /**
    * A private helper used to check if the URL is too long
    */
-  private wmsUrlTooLong(sldBody: string, layer: LayerModel): boolean {
+  public wmsUrlTooLong(sldBody: string, layer: LayerModel): boolean {
     return (
       encodeURIComponent(sldBody).length > Constants.WMSMAXURLGET ||
       this.conf.forceAddLayerViaProxy.includes(layer.id)
     );
   }
-
 
   /**
    * Get WMS 1.3.0 related parameter
@@ -134,7 +133,6 @@ export class CsWMSService {
     return params;
   }
 
-
   /**
    * get wms 1.1.0 related parameter
    * @param layer the WMS layer
@@ -189,101 +187,6 @@ export class CsWMSService {
     }
     return params;
   }
-
-
-  /**
-   * Get the SLD from the URL
-   * @param sldUrl the url containing the SLD
-   * @param usePost use a HTTP POST request
-   * @param onlineResource details of resource
-   * @return an Observable of the HTTP request
-   */
-  private getSldBody(
-    sldUrl: string,
-    usePost: boolean,
-    onlineResource: OnlineResourceModel,
-    param?: any
-  ): Observable<any> {
-    // Pass through any sld_bodys already set
-    if (param && param.sld_body && param.sld_body !== '') {
-      return new Observable(observer => {
-        observer.next(param.sld_body);
-        observer.complete();
-      });
-    }
-    // For ArcGIS mineral tenements layer we can get SLD_BODY parameter locally
-    if (UtilitiesService.isArcGIS(onlineResource) && onlineResource.name === 'MineralTenement') {
-      return new Observable(observer => {
-        param.styles = 'mineralTenementStyle';
-        const sld_body = MinTenemStyleService.getSld(onlineResource.name, param.styles, param.ccProperty);
-        observer.next(sld_body);
-        observer.complete();
-      });
-    }
-
-    // For GeoSciML 4.1 we can get SLD_BODY parameter locally
-    if (onlineResource.name === 'gsmlbh:Borehole') {
-      return new Observable(observer => {
-        param.styles = onlineResource.name;
-        // If borehole name was set in filter
-        let nameFilter = '';
-        if ('optionalFilters' in param && param.optionalFilters.length > 0) {
-          for (const filt of param.optionalFilters) {
-            if (filt.label === 'Name') {
-              nameFilter = filt.value;
-              break;
-            }
-          }
-        }
-        const sld_body = GSML41StyleService.getSld(onlineResource.name, param.styles, nameFilter);
-        observer.next(sld_body);
-        observer.complete();
-      });
-  }
-    // If there is no SLD URL coming from config
-    if (!sldUrl) {
-      return new Observable(observer => {
-        observer.next(null);
-        observer.complete();
-      });
-    }
-
-    let httpParams = Object.getOwnPropertyNames(param).reduce(
-      (p, key1) => p.set(key1, param[key1]),
-      new HttpParams()
-    );
-    httpParams = UtilitiesService.convertObjectToHttpParam(httpParams, param);
-    if (usePost) {
-      return this.http
-        .get(this.env.portalBaseUrl + sldUrl, {
-          responseType: 'text',
-          params: httpParams
-        })
-        .pipe(
-          map(response => {
-            return response;
-          })
-        );
-    } else {
-      return this.http
-        .post(this.env.portalBaseUrl + sldUrl, httpParams.toString(), {
-          headers: new HttpHeaders().set(
-            'Content-Type',
-            'application/x-www-form-urlencoded'
-          ),
-          responseType: 'text'
-        })
-        .pipe(
-          map(response => {
-            return response;
-          }),
-          catchError((error: HttpResponse<any>) => {
-            return observableThrowError(error);
-          })
-        );
-    }
-  }
-
 
   /**
    * Get the NvclFilter from the URL
@@ -341,7 +244,6 @@ export class CsWMSService {
         );
     }
   }
-
 
   /**
    * Get the WMS style URL if proxyStyleUrl is valid
@@ -464,13 +366,13 @@ export class CsWMSService {
       // Set 'usePost' if style request parameters are too long
       const usePost = this.wmsUrlTooLong(this.env.portalBaseUrl + layer.proxyStyleUrl + collatedParam.toString(), layer);
       // Perform request for style data, store subscription so we can cancel if user removes layer
-      this.sldSubscriptions[layer.id].push(this.getSldBody(layer.proxyStyleUrl, usePost, wmsOnlineResource, collatedParam).subscribe(
-        sld_body => {
-          const longResp = this.wmsUrlTooLong(sld_body, layer);
+      this.sldSubscriptions[layer.id].push(
+        this.sldService.getSldBody(layer.proxyStyleUrl, usePost, wmsOnlineResource, collatedParam).subscribe(sldBody => {
+          const longResp = this.wmsUrlTooLong(sldBody, layer);
           // Create parameters for add layer request
           const params = wmsOnlineResource.version.startsWith('1.3')
-            ? this.getWMS1_3_0param(layer, wmsOnlineResource, collatedParam, longResp, sld_body)
-            : this.getWMS1_1param(layer, wmsOnlineResource, collatedParam, longResp, sld_body);
+            ? this.getWMS1_3_0param(layer, wmsOnlineResource, collatedParam, longResp, sldBody)
+            : this.getWMS1_1param(layer, wmsOnlineResource, collatedParam, longResp, sldBody);
 
           let lonlatextent;
           if (wmsOnlineResource.geographicElements.length > 0) {
@@ -490,11 +392,11 @@ export class CsWMSService {
 
           // Perform add layer request
           layer.csLayers.push(this.addCesiumLayer(layer, wmsOnlineResource, params, longResp, lonlatextent));
-          layer.sldBody = sld_body;
+          layer.sldBody = sldBody;
 
           // For 1.3.0 GetFeatureInfo requests need lat,lng swapped to lng,lat if polygon filter present
           if (wmsOnlineResource.version === '1.3.0' && collatedParam.optionalFilters.find(f => f.type === 'OPTIONAL.POLYGONBBOX')) {
-            layer.sldBody130 = this.reverseSldBodyPolygonFilterCoordinates(sld_body);
+            layer.sldBody130 = this.reverseSldBodyPolygonFilterCoordinates(sldBody);
           }
         }));
     }
