+++ projects/portal-core-ui/src/public-api.ts
@@ -21,6 +21,7 @@ export { CsMapObject } from './lib/service/cesium-map/cs-map-object';
 export { CsMapService } from './lib/service/cesium-map/cs-map.service';
 export { ManageStateService } from './lib/service/permanentlink/manage-state.service';
 export { MinTenemStyleService } from './lib/service/style/wms/min-tenem-style.service';
+export { SldService } from './lib/service/style/wms/sld.service';
 export { NotificationService } from './lib/service/toppanel/notification.service';
 export { DownloadWcsService } from './lib/service/wcs/download/download-wcs.service';
 export { CsCSWService } from './lib/service/wcsw/cs-csw.service';
